高并发的写优化

+++ src/main/java/com/example/thumb/sys/entity/Blog.java
@@ -42,7 +42,7 @@ public class Blog {
     /**
      * 点赞数
      */
-    private Integer thumbcount;
+    private Long thumbcount;
 
     /**
      * 创建时间
